Remove reference to WORKER_METRICS inside db_metrics
